5396: rename search list item again
