+++ src/components/card-item-list/card-list-item/card-list-item-cover.tsx
